@@ -18,14 +18,14 @@
     "@reduxjs/toolkit": "^1.9.3",
     "gsap": "3.11.5",
     "leva": "^0.9.34",
-    "next": "13.3.0",
+    "next": "^13.4.4",
     "next-redux-wrapper": "^8.1.0",
     "next-sitemap": "^4.0.7",
     "next-transpile-modules": "^10.0.0",
     "nextjs-google-analytics": "^2.3.3",
     "r3f-perf": "^7.1.2",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
     "react-use": "^17.4.0",
     "three": "^0.152.2",
