@@ -1,12 +1,13 @@
+"use client";
 import { useAppDispatch, useAppSelector, RootState } from "@/libs/store/store";
 import { useEffect, useState } from "react";
-import { useRouter } from "next/router";
+import { usePathname } from "next/navigation";
 import { setMobileMenuOpen } from "@/slices/mobileMenuSlice";
 import MobileNavigation from "@/components/common/mobileMenu/MobileNavigation";
 import CloseMobileMenu from "@/components/common/mobileMenu/CloseMobileMenu";
 
 export default function MobileMenu() {
-  const router = useRouter();
+  const pathname = usePathname();
   const dispatch = useAppDispatch();
   const selector = useAppSelector;
   const mobileMenuOpen = selector((state: RootState) => state.mobileMenu.open); // updated
@@ -32,17 +33,13 @@ export default function MobileMenu() {
     mobileMenuClassHandler(overlayActive);
   }, [overlayActive]);
 
-  // Hook when the route is changing
+  // Triggered every time the path changes
   useEffect(() => {
-    router.events.on("routeChangeStart", () => {
-      dispatch(setMobileMenuOpen(false));
-    });
-  }, [dispatch, router.events]);
+    dispatch(setMobileMenuOpen(false));
+  }, [dispatch, pathname]);
 
   return (
-    <header
-      className={`fixed right-0 top-0 z-30 h-screen w-[300px] bg-mirage transition-all duration-200 ` + mobileMenuClass}
-    >
+    <header className={`fixed right-0 top-0 z-30 h-screen w-[300px] bg-mirage transition-all duration-200 ` + mobileMenuClass}>
       <div className="flex h-full w-full flex-col">
         <div className="flex items-center justify-between border-b-2 border-bleached px-30px py-10px pb-5px">
           <div className="py-10px xl:py-20px">Logo Here</div>
