@@ -1,21 +1,18 @@
+"use client";
 import React from "react";
-import { GoogleAnalytics } from "nextjs-google-analytics";
-import "@/styles/globals.scss";
 import { Provider } from "react-redux";
 import { store } from "@/libs/store/store";
+import { GoogleAnalytics } from "nextjs-google-analytics";
 import fonts from "@/libs/constants/fonts";
 import LoadingScreen from "@/components/common/loading/LoadingScreen";
 import LevaControls from "@/components/three/containers/LevaControls";
-import ThreeCanvas from "@/components/three/ThreeCanvas";
-// import dynamic from "next/dynamic";
-// const DynamicThreeCanvas = dynamic(() => import("@/components/three/ThreeCanvas"), {
-//   loading: () => null,
-//   ssr: false,
-// });
-
-export default function App({ Component, ...props }) {
-  const { pageProps } = props;
-
+// import ThreeCanvas from "@/components/three/ThreeCanvas";
+import dynamic from "next/dynamic";
+const Canvas = dynamic(() => import("@/components/three/ThreeCanvas"), {
+  loading: () => null,
+  ssr: false,
+});
+export default function Wrapper({ children }: { children: React.ReactNode }) {
   return (
     <Provider store={store}>
       <GoogleAnalytics trackPageViews />
@@ -23,9 +20,9 @@ export default function App({ Component, ...props }) {
         <div className="grain" />
         <LoadingScreen />
         <LevaControls />
-        <Component {...pageProps} />
+        {children}
         <div className="absolute left-0 top-0 z-1 h-screen w-full">
-          <ThreeCanvas />
+          <Canvas />
         </div>
       </div>
     </Provider>
