@@ -1,3 +1,4 @@
+"use client";
 import { Mesh } from "three";
 import React, { useEffect, useRef } from "react";
 import { useFrame } from "@react-three/fiber";
@@ -6,7 +7,7 @@ import { setSceneObjectsLoaded } from "@/slices/threeSlice";
 import { ThreeStateObjectsLoadedAction } from "@/types/three/state";
 
 const Box = ({ color = 0xf5d9d4, stateScene, name, ...props }) => {
-  const ref = useRef<Mesh>();
+  const ref = useRef<Mesh>(null);
 
   // Set up the Redux State dispatch
   const dispatch = useAppDispatch();
@@ -26,9 +27,13 @@ const Box = ({ color = 0xf5d9d4, stateScene, name, ...props }) => {
   }, [dispatch, name, stateScene, ref]);
 
   useFrame((state, delta) => {
-    ref.current.position.y = 10 + Math.sin(state.clock.elapsedTime) * 3;
+    if (ref.current) {
+      ref.current.position.y = 10 + Math.sin(state.clock.elapsedTime) * 3;
+    }
+
     // ref.current.rotation.x = ref.current.rotation.y = ref.current.rotation.z += delta;
   });
+
   return (
     <mesh ref={ref} scale={10} position={[0, 0, 120]} {...props}>
       <boxGeometry />
