@@ -1,3 +1,4 @@
+"use client";
 import { Plane } from "@react-three/drei";
 import GroundGUI from "./GroundGUI";
 import groundSettings, { GroundSettings } from "./groundSettings";
@@ -7,13 +8,16 @@ const Ground = ({ showGUI = false }) => {
   const settings: GroundSettings = showGUI ? GroundGUI() : groundSettings;
 
   return (
-    <Plane
-      args={[settings.geometry.x, settings.geometry.y]}
-      rotation={[settings.rotation.x, settings.rotation.y, settings.rotation.z]}
-      position={[settings.position.x, settings.position.y, settings.position.z]}
-    >
-      <meshStandardMaterial color={settings.color} />
-    </Plane>
+    settings.rotation &&
+    settings.position && (
+      <Plane
+        args={[settings.geometry.x, settings.geometry.y]}
+        rotation={[settings.rotation.x, settings.rotation.y, settings.rotation.z]}
+        position={[settings.position.x, settings.position.y, settings.position.z]}
+      >
+        <meshStandardMaterial color={settings.color} />
+      </Plane>
+    )
   );
 };
 
